Centralize MD5 hashing utility
