+++ app/src/main/kotlin/com/d4rk/cleaner/app/clean/scanner/utils/helpers/DuplicateUtils.kt
@@ -1,7 +1,7 @@
 package com.d4rk.cleaner.app.clean.scanner.utils.helpers
 
 import java.io.File
-import java.security.MessageDigest
+import com.d4rk.cleaner.core.utils.extensions.md5
 
 /**
  * Groups duplicate files by their original version.
@@ -25,16 +25,3 @@ fun groupDuplicatesByOriginal(files: List<File>): List<List<File>> {
     }
     return groups
 }
-
-private fun File.md5(): String? = runCatching {
-    val buffer = ByteArray(DEFAULT_BUFFER_SIZE)
-    val md = MessageDigest.getInstance("MD5")
-    inputStream().use { stream ->
-        var read = stream.read(buffer)
-        while (read > 0) {
-            md.update(buffer, 0, read)
-            read = stream.read(buffer)
-        }
-    }
-    md.digest().joinToString("") { "%02x".format(it) }
-}.getOrNull()
